feat: export plaintext and HTML fields values
This is part of story #23469: export reports with linked artifact in xlsx format

Text field values (either in HTML or in plaintext) are now exported in
the XLSX. Some escaping is done for HTML content.
Rows heights are modified to fit the multilines content that can comes
from HTML or plaintext content.

Change-Id: If3c2057fcf3e1e99eefde66e81a2ace1b130f1a1