chore: Reorganize namespaces
part of story #24969 [modal] add and remove links without types

No functional change expected.

ChangesetValue does not have to be under Changeset namespace, it is now
alongside it. Also, this splits REST (with payloads and stuff) from
"domain" objects.

"Link" is renamed to "ForwardLink" so that we can write a "ReverseLink"
interface in another change. We should distinguish the two, they have
very different meaning and are handled differently.

Change-Id: I5e63f089e7bd566fdb178f04834a1610e9faa7d1