feat: Use @floating-ui for TLP dropdowns
Positioning floating element is hard to do with pure CSS and we should
trust dedicated JS library to do the job, like it was done for popovers.

Floating-ui is the next generation of Popper (the latter is still used for TLP
popovers but will be upgraded in a dedicated contribution):
https://floating-ui.com/docs/motivation

No functional changes expected everywhere TLP dropdowns are used.

Part of story #24197: search for Documents with criteria

Change-Id: I6921b0b1b1f740def0e88d6212e06c69135a55fe