chore: Initialize XLSX lib
This is part of story #23469: export reports with linked artifact in xlsx format

This is the first step for a library that will contain all the general
types and methods for XLSX generation.
For now, it contains only Cell types, and is only used in cross-tracker
report file.

No functional change expected.

Change-Id: I4706df4e6e7451c8033724ab8ae0acd927d2350b