Fetch reports of the current tracker into the properties of the cross-report modal
This is needed to be able to quickly list the reports of the first
export level. For the others levels we will be forced to fetch them
through the REST API.

The selector is not yet working but you can check that all the values
are available in the in `data-properties` attribute of the mount point.

Part of story #23469: export reports with linked artifact in xlsx format

Change-Id: I7a1451b1d07e4405ec61f1ce8dcc0178b8670931