Retrieve (and store) Nix derivations from a Cachix cache
This will avoid rebuiding the Nix derivations everytime to save some CPU
time.

The script tools/rpm/build_rpm_inside_container.sh can be used to push
the derivations to the cache. It will only attempt to do it when a file
/cachix_auth_token is set inside the container with some content.

A cache named `tuleap-community` has been created on Cachix.org. Right
now it is empty so no improvement will be visible. The pipeline building
the packages will be modified to set the /cachix_auth_token file.

Part of request #26358: Cache the results of the Nix derivations used when building the packages

Change-Id: I2fe97d21c26d69b20f755c70dc25e4edb1ca71b9