feat(forge-config): Default value from attributes (plugins)
After Core, plugins can now define default values for their
configuration variables. It's now the preferred method over setting the
value in DB directly.

This part is a bit trickier than for Core because we have to keep the
lazy loading aspect of Plugins (otherwise the performance will be
severly hit). The design decision was to resolve the default values when
creating the cache vs only store classes that holds configuration
variables because if we chose the later, when loaded from cache the
classes cannot be loaded (plugins not autoload).

It's also a bit trickier than for other loaders because due to the way
ForgeConfig works, the customized values, store in database, will be
loaded **before** the attributes. Since the load of
ForgeConfig::loadInSequence is done at the earliest in the request
processing AND that variables that comes from the DB are loaded during
loadInSequence, then if a variable is already set it has a value in DB.
Due to lazy loading, the value that might come from attributes will be
loaded after (when plugin system is loaded) so existing values **must
not** be override.

Finally, I propose to share a common interface between GetKeysList event
and SerializedPluginProxy to gather classes that exposes ConfigKey. It's
handy but I don't know if it's a good idea.

How to test ?

 tuleap config-get feature_flag_doc_gen_cross_tracker_report_xslx_export

now returns 0 instead of saying that value is not defined

Part of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: Idbd3614f94659bfae7c502ef97ea3826ba21aad5