Cache all paths produced when building Nix derivations
This can be useful when doing some changes on derivations requiring to
build another derivation e.g. the tuleap-git-bin package where we have
to build the static Git before building the RPM.

Since it does not change much when uploading the derivation the first
time and it can be nice to have when modifying our derivations.

Part of request #26358: Cache the results of the Nix derivations used when building the packages

Change-Id: Ie4ee227b2e26abd87474d6b074940e8d0c1b31df