Artifact link types to use can be selected for the first level
Only the artifacit link types used in the tracker can be selected.
Since this is not yet used in the export the easiest way to make sure
the selector work as expected is to use the Vue dev tools to observe the
values inside the components.

Part of story #23469: export reports with linked artifact in xlsx format

Change-Id: Ifbf60c23cf0e811633e617d9bb2ff8a46b66fbe9