`cachix watch-exec` conflicts with `nix-build` options
Two mistakes were made in 340de830e37a56bc60a228beeb5eed9160e8bee0:
* the name of the cache to push to was forgotten
* the options we use with `nix-build` conflicts with `cachix`

As a result some packages are not built in the build pipeline.
To avoid more surprises, the call to `nix-build` is moved to a dedicated
script.

Part of request #26358: Cache the results of the Nix derivations used when building the packages

Change-Id: I02a49ff12b99327321336eb0ba8f8a76c773aa08