fix(forge-config): Propagate config for TULEAP_SYS_DBPORT
With refactorings of ForgeConfig loading we lost the ability to customize
the database with environment variables. This commit brings it back.

This commit also fix an issue with mediawiki that was not possible to
run with a database not running on 3306 port.

In addition to that:
* The legacy `TULEAP_DB_SSL_MODE` and `TULEAP_DB_SSL_CA` env variables
  are removed in favor of TULEAP_SYS_ENABLESSL, TULEAP_SYS_DB_SSL_CA and
  TULEAP_SYS_DB_SSL_VERIFY_CERT
* the --skip-database option was removed because it was there only for
  tests and the error was introduced because of this option (it skipped
  a part of the code that was triggering the initial issue).

Closes request #26380 mediawiki doesn't work when db port is not 3306
Part of: story #22708 flexibility in the management of Tuleap Configuration

Change-Id: Iaa7349ad0aaf73557c34aafdb513cb7ccb0cabee