feat: Display errors in tlp-modal-feedback
part of story #24969 [modal] add and remove links (replace current
artifact link)

How to test:
When there is an error while retrieving the linked artifacts (either GET
link types or GET linked_artifacts), an error message will be shown in
the modal feedback (that is fixed and does not scroll) instead of the
links field itself.

You can test it by throwing a RestException in ArtifactsResource.

When creating a new sub-item (for example a Task on a User Story), the
modal feedback still shows its future parent.

Notes:
We should do the same for other fetch operations but it should come
later.

Change-Id: I68d511c7016063f5e84ce820ddc383c5deaa205d