chore: Initialize the organization of the reports data
This is part of story #23469: export reports with linked artifact in xlsx format

No functional changes excpected here.

This commit introduces a new intermediate format to store Artfiacts
data. For now, it stores only the ArtifactRepresentation in a map.
In the next commits, it will help to store artifacts from the second report
and the links between the artifacts.

Change-Id: Ibef4d94ab6da5a3f679d4a749444ac821cdeded0