refactor: Display parent retrieval error in tlp-modal-feedback
part of story #24969 [modal] add and remove links (replace current
artifact link)

How to test:
When there is an error while retrieving the future parent of an
artifact, an error message will be shown in the modal feedback instead
of at the top of the scrolled modal.

You can test it by throwing a RestException in ArtifactsResource and
trying to create a new sub-item (for example a Task on a User Story).

Change-Id: Ic49b6c7fa87223d7eefa7a22cc22428891c6d496