feat: Begin to display linked artifacts from selected level 2
This is part of story #23469: export reports with linked artifact in xlsx format

To test this feature, you will have to add an URL parameter to provide
the expected level 2 report id: export_report_id_2

If this parameter is provided, then linked artifact filtered with select
level 2 report are now added at the end of the generated XLSX.

Neither level 2 headers nore duplication of rows are done for now.
This will be added in next commits.

Change-Id: Iaa0fb84d341302f1082d1e4932a4d917e60f07c5