feat: Restrict allowed types to "no type" and Child (forward)
part of story #24969 [modal] add and remove links (replace current
artifact link)

In the current story, only "no type" and Child (forward link) are
allowed to be chosen. The rest of the types will be allowed in the next
story (story #24970 [modal] select type for links)

Notes:
I know filtering in the template is crappy but this should be temporary.

Change-Id: Ia74ce69174eb1f7de4a2548c94955b5069025065