refactor: Give only controller to LinkField
part of story #24969 [modal] add and remove links (replace current
artifact link)

No functional change expected.

Notes: Giving both properties from "field" and a presenter from a
controller was confusing. It created two equivalent "sources of truth".
The custom element should have only one source of truth: its controller
that builds presenters for it.
With this contribution, the controller has two methods and builds two
distinct presenters: one for the field, another for the currently linked
artifacts. I first tried this with a single presenter, but it turns out
that it was way too complicated, it created several problems. This is
simpler.

Also, since ArtifactCrossReference is a composite (it is built from a
TrackerShortName and an Artifact Identifier), there is no need to proxy
it. It can be fully in the Domain.

Change-Id: I78188a9413af49921713fc6ddfcef394177f7e94