ListPicker updates its state accordingly to the source select disabled attribute
With this change it is now possible to disable an active select (or the
other way around) when ListPicker is used. Until now only the state when
initializing ListPicker was taken into account.

This change is needed for the export configuration modal so we can
activate fields once the user has selected other required fields.

The easiest way to test this change is to play around in the tlp-doc
page.

Part of story #23469: export reports with linked artifact in xlsx format

Change-Id: I590019c0f6199bb65bdcf1446ec6f6b64778ae8e