feat: Export selected level 3 headers
This is part of story #23469: export reports with linked artifact in xlsx format

Tracker name with merges and field labels must be displayed in the
generated XLSX document.

To have 3 levels of link, you can choose to export Release -> Sprints ->
User stories in a project that uses the Agiledashbord Scrum service.

Change-Id: Iab3262a56253c7968d027198dadf4295d13550c9