refactor: DRY filename preview property
As of today the filename preview is displayed by two different
components that:
* detect if the preview can be displayed
* compute the preview
* display the preview with additional info (tooltip)

While what the preview be like is specific (new version vs new file),
how and when to display it is common to both, it can be extracted to
avoid to repeat ourselves too much.

No functional changes.

Part of story #24192 preview file name at upload

Change-Id: Ide6bcf6eb97caa9cf031c96b4cb14ebe9a76b0e4