feat: Configure criteria & columns
The admin can decide which properties can be used in search criteria and
search results columns.

If administrator does not change anything, then all properties are used
as criteria/columns.

Only add and removal is supported. The order is not taken into account,
it will be part of a dedicated contribution.

Part of story #24197: search for Documents with criteria

Change-Id: I4be0a16dd4adb2fdda6e0822f8a289d07cf46fc6