Remove @each loop for cross-reference badges
part of request #26395 Avoid @each rules for basic swatch color usage

No functional change expected. Cross-reference badges are still colored
by the tracker color. They still display correctly on dark background in
the switch-to modal. TTM "current bugs or requirements" are always
white, unless there are more than two (in a dropdown).

Notes:
This also removes the duplication of the styles between FP and BP. The
badges' SCSS is moved to a dedicated lib that is shared between them.

Change-Id: Ic4284258d9031539c898744c9f08a6d574e60618