refactor: split ParentArtifact type from Artifact
part of story #24969 [modal] add and remove links (replace current
artifact link)

No functional change expected: When you create a new sub-item (for
example a new Story in an Epic in Planning view), the modal will keep
fetching the parent and show an "alert" at the top like "The artifact
will be linked to <title>".

This refactoring aims to delete the generic "Artifact" type. We should
use specialized types.

Change-Id: If7d4c186928418565245096898a391bc511ff24f