request #26400 Jira project import should not stop when artifact link type creation fails
As jira import can be done without types we should move forward. Import
is something in which some informations will be lost, at this stage,
it's better to move forward and raise a warning to see if it can be
fixed later on (either tidy up the data at Jira level or fix import
process at Tuleap level).

Change-Id: I516ffe08d5e4b8f275e0099193c4565ac8365e6e