feat: toggle add link button disabled
part of story #24969 [modal] add and remove links (replace current
artifact link)

When you type a number matching an artifact id in the new link field,
you can select it (by click or with keyboard). When you select it, the
"Add link" button will be enabled. It does nothing for now.
When you clear the selection, the "Add link" button will be disabled
again.

Notes:
Link selector does not create <options> anymore. It is still bound on a
<select> element for now, but only for legacy reasons.

Change-Id: If94875f77bca80df2b3dcae1558f1442e7ee4f28