refactor: fix proxy naming
No functional change expected in new artifact link field.

Notes:
TrackerShortnameProxy is moved from Caller adapter to REST adapter.
Caller adapter is meant to isolate the arguments passed to the
createModal() function. The Tracker shortname goes through the modal's
controller, but originally it comes from REST, so its proxy should be in
REST adapter.
LinkableArtifactIdentifier is renamed to LinkableNumber. We don't know
at this step if it's really an Artifact id, we just know that it looks
like a number and is not the current artifact id.

Change-Id: I3b309fea7f0d56e139ab3a5dc6bf024019f30265