request #26408 Import of jira issue link types with not supported characters
I wanted to re-use the regexp defined in TypeValidator but it seems that
Cocur/Sluggify does a char/char conversion so it doesn't work out of the
box. Nevertheless, the goal is not to have a perfect tranformation in
every possible case but to support most common conversion scenario
(unsupported char in the middle of the label).

This patch also make it clearer the dependencies between the various
places where link types are manipulated: when they are created in bulk,
when they are "faked" to re-create a missing information (Epic/Issues)
and when the artifacts are finally generated.

Change-Id: I0dd207a76130572228575ce3c458ae4e2438847d