feat: 3rd party integration knows about references
This exposes the list of supported keywords for a given project.

This commit also add support of `project` for goto page so integrations
can re-use the given argument (bug #gpig:1234) literally for
constructing the goto url instead of having to resolve the shortname
first.

Part of story #25734 have Tuleap cross references in mediawiki pages

Change-Id: Ia08e0201744abe779923777e0fa9ff2a2a207644