Retrieve Mediawiki extensions and Tuleap skin during the "prepare" phase
The idea is to retrieve the Tuleap skin, MW extensions and their
dependencies in the same way we retrieve dependencies for Tuleap.
To build the final "MediaWiki Tuleap flavor" package we will take the
resulting content generate by `composer install` and merge it with the
release tarball of Mediawiki.

Adding or updating extensions/skins can be done with standard Composer
commands.

At this stage there is not really something to test besides having
content with the MW structure retrieved when running `make composer`.

pnpm workspaces have been updated to only include the packages we need.
We cannot exclude all packages located inside **/additional-packages/**
directories because Turborepo does not currecntly respect the exclusion
[0].

Part of story #26042: install mediawiki standalone

[0] https://github.com/vercel/turborepo/issues/998

Change-Id: I4517bd2f6e5730e826494b1e3894d5387d3c7c40