Build a "MediaWiki Tuleap flavor" RPM
This RPM is built from 3 distinct parts:
* the MediaWiki release tarball
* the MediaWiki extensions and skins needed for the Tuleap usage
  retrieved with Composer
* MW settings and overrides needed to use with Tuleap

Each part is retrieved and built separately and modified a bit to avid
conflicts and then put together and cleaned.

The main modifications to avoid conflicts is to move the autoloaders of
the MediaWiki and the MediaWiki extensions and skins packages on
different paths. A manually crafted `vendor/autoload.php` file loading
the two autoloader is deployed.

Note that we do not pull guzzlehttp/guzzle with the extensions/skin we
retrieve with Composer since the release tarball of Mediawiki already
comes with it.

At upgrades if new conflicts arise we cannot miss them since building
the derivation will throw an error. If guzzlehttp/guzzle is no more
provided by the release tarball of Mediawiki it will also be hard to
miss since the OAuth2 flow between Tuleap and MW will be broken.

The RPM puts the MediaWiki files under `/usr/share/mediawiki-tuleap-flavor/`
to be consistent with the other files we deploy.

This contribution of this plugin also enables the build of the RPM of
the Tuleap plugin so we can start setting the dependencies between the
different pieces.

At this stage it is still a shot in the dark since we cannot test it yet.
Future contributions will deal the deployment of the required nginx and PHP-FPM
configurations. We can however look at the field deployed by the RPM to
check if they look coherent. Adjustments will be made if needed.

Part of story #26042: install mediawiki standalone

Change-Id: I4960e0aa76b4a02b0cd4632f5974951e997a79f9