request #26758: Error is thrown in Chrome when an autofill event is encountered
The issue is caused by a bug in @vueuse/core.
To fix it, we upgrade @vueuse/core from 7.5.3 to 8.4.2.
The only version changing the code we use in the library is the 7.6.2,
see corresponding changelog: https://github.com/vueuse/vueuse/releases/tag/v7.6.2

No functional changes are expected, it should still be possible to
select a service in the sidebar using the keyboard shortcuts.

Change-Id: I08aba7e99035efec703955a59e620ddaea530fbc