feat: Deal with admin permissions
At this stage (this story) we covers only default permissions. The
configuration part (where we define who is writer, sysop & co) will be
done in the next story.

This commit introduce a dependency between mediawiki_standalone and
legacy mediawiki plugin because we need the definition of a Global
Mediawiki Administrator.

The REST test will be added when the functional part is completed.

Part of: story #25738 manage Tuleap users & project access

Change-Id: If6c96f9a2147df90ec7ffd802374cd110daa17b4