chore: REST API types for Selectbox field structure
No functional change. Typechecking should not raise errors.

I chose to duplicate the big field structure union type in docgen-docx
lib so that we can restrict field types to only what is needed. Reusing
is good, unless it forces us to do a bunch of extra work every time
there is an update. To prevent this, we redefine types by picking only
the handful of attributes that are actually needed for docgen-docx. That
way, this lib stays as it is, and we are free to add more details to the
REST API types.

part of: request #22648 Convert AngularJS directives and controllers to
hybrids web components

Change-Id: Ic7499e1aaf3fd6b9adcb6a6e91b443cc978c1c59