feat: disallow linking twice the same artifact
part of story #24969 [modal] add and remove links (replace current
artifact link)

How to test:
given you added a new link to an artifact
- when you type its ID in the link-selector
- when you switch to "Parent" link type and browse the possible parents
- the link selector item should be disabled and greyed out. You cannot
select it again.

Same for an existing link to an artifact (that is not marked for
removal).

Notes:
The "(already linked)" mention is not yet added to the item's label. It
will require another contribution.

Change-Id: I89392c50cd44b37d7fa34637702479072ec0d4dc