request #26789 Tracker report : wrong result when we search on date > given_date
For date field without time, it gives the wrong result when we search on date > given_date.

When we search for a date > at a given_date, we add a day at the given_date and search dates > at give_date + 1d.
If this is a date field without time, the date is saved at midnight, so the comparison will return all artifacts with give_date + 1d + 1s

It work well for expert queries

Change-Id: I96b2c6d2ddff4a5e0c3d8992c4b0e4c371167523