feat: Do not take "mark for removal" into account for link checks
part of story #24969 [modal] add and remove links (replace current
artifact link)

How to test:
given you have an existing link to an artifact A (or you add a new link
to A)
- when you type its ID in the link-selector
- when you switch to "Parent" link type and browse the possible parents
- the link selector item should be disabled and greyed out. Even if you
mark the link for removal, it stays disabled.

when an artifact already has a parent (or you add a new parent to it),
the "Parent" link type is disabled. It is re-enabled if you remove the
new parent link (existing link type cannot be marked for removal yet).

Change-Id: Ia13b9c64663c135c9f1dbbcb3bb23bb257969546