Fix request #26793: TOS should respect user's language
If user has fr_FR as language in user prefs, then in project creation,
the modal should display terms of services in french.

Default global $Language does not know the locale by default. A first
call to `$Language->getText` must be call to load the locale based on
the current user.
The solution here is to use the Language provided by the current user
(who knows its prefered locale).

Change-Id: Ic36904506a6dc60310e62eb6f450da93ed8d3e05