Package the mathoid CLI
We package the mathoid CLI using vercel/pkg [0] like we do for tuleap-realtime.
This allow us to use whatever version of Node we want.

The config file of Mathoid is not expected to be modified on the servers.
For now it is based on the "production" config file provided in the Mathoid repository.

We use the current latest commit of the mathoid repository instead of the last tagged
version because some fixes are missing.

To test you can try to build the package and deploy it on your dev instance. You should
be able to play with the Mathoid CLI binary.

Part of story #26042: install mediawiki standalone
Closes request #25725: Mathoid node app cannot find the relevant node bin

[0] https://github.com/vercel/pkg

Change-Id: I3f1d74f91a9f84cc74d0810e4ea2e3e232ae8132