feat(mediawiki): define readers
Readers are those who can access the project.

There are a few things in User & Project builders that are added in this
commit but not used. It's artifact of the first implementation that
proved to be useless. Nevertheless, I think they are useful builders so
I left them here.

Part of: story #25738 manage Tuleap users & project access

Change-Id: I078ad25a748cad5a6d5d69f470bd060d9f3553cb