Generate a LocalSettings file when installing the mediawiki_standalone plugin
The file can also be generated via an hidden command
`tuleap mediawiki_standalone:generate_local_settings` to ease the tests
(and make it easy to regenerate it if it becomes broken).

This contribution does not yet take the creation of the OAuth2 app into
account.

Part of story #26042: install mediawiki standalone

Change-Id: I0fe36dc07510b8bc3179967fc44b03ccc0234e21