feat: Add artifact badge in parent note
part of story #24969 [modal] add and remove links (replace current
artifact link)

Since we have to display some HTML in a sentence, we have no choice to
split it in two. It cannot be embedded inside the translated string
because Hybrids template tag "html" escapes the html tags for security
reason.

How to test:
- Open the artifact modal in creation mode
--> A generic note is displayed

- Open it in edition mode
--> The note contains the currently edited artifact cross reference in a
badge.

Change-Id: Ib7b3e5efbd07933f7fd934332c5fd618afcbe06c