feat(mediawiki): Create new mediawiki instance
The feature is not yet connected to the interface, to be able to test it
you will need to send the creation payload manually:

    tuleap worker:enqueue tuleap.mediawiki-standalone.instance-creation '{"project_id":101}'

For this to work you will also need to adjust the farm admin password in
InstanceCreationWorkerTask. It's hardcoded as of today because it will
be removed from TuleapFarm MW extension soon so I don't want to
introduce something that will be removed in couple of days.

You will also need to define the pre-shared key at Tuleap side (it's
already being generated for MW but Tuleap doesn't store it yet). You can
do it with

    tuleap config-set mediawiki_standalone_shared_secret ...

This should be done automatically at some point but as this part of the
code is currently being worked on in another patch, I prefer to avoid
conflicts.

Part of: story #24228 create/update/delete mediawiki-standalone in my project

Change-Id: I78996c34dd60205234cee055867db6562a3e2658