feat(mediawiki): Suspend & Resume instances
As the implementation starts to be a lot of copy/paste this is mainly a
refactoring of the previous implementation to mutualize the work.

I didn't mutualize everything yet because I want to see how it fits with
actual service activation / deactivation before.

Nix packaging evolves because there is a need for a specific
configuration file when an instance is suspended. Extract of the
information from the not public design document:

> Suspending an instance will now keep the instance available, but will
> limit permissions only to sysop group (others get no permission). This
> is achieved by loading a static file on suspended instances. This file
> should be located in wiki root and be called
> `LocalSettings.SUSPENDED.php`.
> Once the file is in place, loading will happen automatically.

Part of: story #24228 create/update/delete mediawiki-standalone in my project

Change-Id: Ie35cc2ba960c2606b6cb63b07d36356f114b6968