Do not pull the glibc from nixpkgs when running macOS
It cannot work. Note this means that the binary build by the
@tuleap/plugin-mediawiki_standalone-mathoid package when running on
macOS is unlikely to work correctly inside the Tuleap Web container.

Given the fact I/we cannot properly develop in this context we will have
to live with it. macOS users wanting to build the RPM will have to use
the same Docker image the CI is using.

Part of story #26042: install mediawiki standalone

Change-Id: I1e98cab70e9588cf0c73230434943c188751a238