feat: allow admins to activate a plugin not in db
Until now, in order to be displayed in the "Services" list in project
administration, plugins had to be first injected inside the database for
each and every service. This was usually done with `install.sql` and
copy/paste of an INSERT INTO ... SELECT query.

The, as services are duplicated at SQL level when project are created it
was working just fine.

But since we introduced the XML based project templates it no longer
works that fine because XML based templates no longer clone a list of
services. They are copied from the XML declaration of services. Hence
all templates must list the list of all possible services, even disabled
because if they are not there at project creation, they won't appear.

This is odd and it's time to move forward: Service activation shouldn't
depend from the database but must be managed by the Plugin itself.

This commit is a first implementation and should be self sufficent.
However the management of services have grown mad and with this new
event, a plugin is expected to implement 5 hooks to "just" have a
service in the sidebar. This will be the purpose of the following
commits.

Part of: story #24228 create/update/delete mediawiki-standalone in my project

Change-Id: I2767e329176717205b98eaf4f4e8cd4f7fe14bb2