Fix request #27131 Updating invitations generates an error
Was introduced by d6548abeb4b86f6e7db258307154677b34d5f22d

Previously the code was expected to return something if update was wrong
but today it will trigger an exception that will be caught by the catch
all of exceptions. As an exception will only be thrown if there is an
error that prevent to write in the database, it's safe to let the catch
all do its job.

Change-Id: I023bf049b9992adbf209f595c23a961a601b6a56