chore: Upgrade GitLab version deployed by Docker to 14.10.3-ce
This is part of story #26801 create a merge request from artifact

We upgrade the GitLab version deployed by Tuleap to the last 14.10 version.

If you never used the GitLab deployed, we can use this container as is.

If you previoulsy used the GitLab deployed and you don't care about the data
stored, you have to remove all the images, containers and volumes
related to GitLab before using this new version.

If you previoulsy used the GitLab deployed and want to keep your data,
you have to follow the upgrade path[1]
To do this, you have to:

* Edit tools/docker/gitlab/Dockerfile to the version needed by the
  upgrade path
* Remove the previous container (docker rm tuleap_gitlab_1)
* build the new tuleap_gitlab image (docker-compose up --build -d gitlab)
* run the new image in a new container in order to have the upgrade
  running (make start-gitlab)

With our last GitLab version deployed you have to follow the following
upgrade path before running this 14.10.3 version: 13.12.15 -> 14.0.12 -> 14.9.0

[1] https://docs.gitlab.com/ee/update/index.html#upgrading-to-a-new-major-version

Change-Id: I3861a5567aae264cd98313a0874d0af91db0a904