chore: Remove legacy project data export
Nobody complained about the deprecation (warning in the interface and in
the [deprecation guide][0]). It's time to clean up.

Now the project data export deals only with XML project structure export.
Since this feature is hidden behind a feature flag, the menu entry is
only displayed if the feature flag is activated.

BTW the feature flag is not anymore based on URL parameter.

Part of request #24983: Removal of legacy project data export

[0]: https://github.com/Enalean/tuleap-documentation-en/blame/13.9/languages/en/deploy.rst#L47

Change-Id: Id691e388a9d65c24426e872bd84bc58d8710c58f