fix: Description in PostMergeRequest can have null value
This is part of story #26801 create a merge request from artifact

A merge request in GitLab can have an empty description that can be
returned as null in the REST API response.

Technically, the isset method in PHP returns false if the key is in the array but
the value is null. So we have to use the method array_key_exists instead.

Change-Id: Ia5ee8fe8d93d4d407c54241eb4cc1586df6496a8